filediff: Check if chunk being highlighted is still valid (bgo#768298)
As described in the bug report, in some situations we can get a chunk
modification event (e.g., from a replacement) followed by a chunk
deletion event (e.g., because the chunks are now equal on both sides).
This messes with us because the deletion will clear the chunk
highlighting immediately, but when the modification results come
through asynchronously we just blindly add the highlighting, even
though this is no longer a change chunk.

With this patch, we check that the chunk we think we're highlighting
still exists before doing anything.