patchdialog: Fix patch creation for missing newline (bgo#770336)
The patch formatted by our patch dialog ends up weirdly joining
together the two parts of a changed chunk if the left hand side of the
chunk doesn't end in a newline. This will only be the case if the chunk
in question covers the last line of a file, and the file doesn't end in
a newline.

The workaround here is more annoying than it should have to be, because
Python doesn't expose what it considers to be a newline character
anywhere.