bin/meld: Setup glib log handler forwarding to Python logging
With this change, we hand warning-and-above logs from GTK+ on to our
Python logger, which gets us some consistency of formatting and
destination, but most importantly means that we also get to silence
GTK+ warnings based on whether the running Meld version is stable or
not.

Effectively, this should silence a lot of noisy complaints about
missing allocations and size requests not happening where GTK+'s
expectations have changed and we don't meet them... or the expectations
aren't actually documented anywhere.