matchers: Copy the passed-in text for mutability and speed
The mutability argument here is pretty clear: we should take a copy of
the sequences, because we can't guarantee that they're not going to
change while we're running our comparison. I'm pretty sure that our
yield points actually happen to guarantee this anyway, but I'd much
prefer being explicit here.

The speed argument is much weirder and more annoying. What this differ
almost always gets passed is a pair of MeldBufferLines instances, which
expose a Python-list-like interface over the lines of a GtkTextBuffer.
What this means in practice is that doing things like iterating over
MeldBufferLines results in half a dozen GTK+ API calls to e.g., get
the text iterator for a visual line, get the start and end of the line,
get the text from that line, clean it up... it's a nightmare and it's
super, super slow.

Doing the whole-buffer copy here does all of this, but only once.
Obviously we pay the memory penalty of copying the whole file, but
given the performance improvements I'm willing to take this as a peak
usage cost.