gutterrendererchunk: Handle more StyleContext ABI breaks (bgo#774810)
Because of the CSS node breakages in GTK+ 3.20, we need extra handling
for doing our new button style drawing in the linkmap.

ui.gtkcompat is a new module with code ported from the GTK+
foreigndrawing demo that handles some context creation. Not all of
the existing code there is necessary, but some of it would let us
use CSS instead of doing manual padding calculations.