dirdiff: Reduce useless display of newer emblem and fix multiple newest
Prior to this commit, we would always display a small star to indicate
when the file/folder in a column was the newest in the comparison. This
mostly makes sense for e.g., modified files, but makes no sense in a
couple of scenarios:

 * We'd show this when there was only one file in the comparison, for
   example when we were in a two-way comparison and the file was
   missing on the other side. Clearly this is pointless, since that was
   already marked as new, so there's no additional information.

 * We'd show the star on the first column whenever both columns had the
   same mtime, which is very misleading.

This handles both of these cases, and also handles three-way
comparisons correctly, so that if two of the three columns have the
same mtime, we'll actually show that *both* of those are newer than the
other.