filediff: Fix chunk deletion for CRLF endings (bgo#778856)
The problem here was that we were always only going backwards by a
single character for the last line in a file, even when file endings
were CRLF. The line ending behaviour here is weird, because of how
Gtk.TextBuffer handles the last line in a file, but it appears to be
correct.

Also added a test for this case.