filediff: Use cursor position for EOF chunk deletion (bgo#778856 take 2)
A problem with the previous approach, demonstrated by the test cases
added, is that GtkTextView inexplicably doesn't bother to add newlines
that match the existing ones while editing a file.

The new approach is to avoid doing this calculation manually ourselves,
and instead rely on cursor position manipulation, since that
will/should always result in us ending up on the last non-linebreak
character of the previous line.