meldwindow: Work around broken UIManager life cycle again (bgo#779880)
This is very similar to bgo#755407, and the "fix" is basically the same,
even though the reproduction is different. This looks like it was
reintroduced when we removed the tabs menu, because the code that was
previously calling ensure_update() was removed along with the tab menu
recreation.

This was fine in most cases, but resurfaced with closing the second of
multiple windows. I have no desire to find out why, and there's no
chance of getting this fixed in GTK+, thus this hack.