filediff: Place cursor at the start of a replaced chunk
This is an ergonomics improvement. In most cases this won't matter at
all. It's sort of more intuitive to keep doing what we used to do, and
place the cursor after the chunk that's just had the action done to
it... after all that's what would happen if you'd just pasted the chunk.

However, it doesn't really feel right to me. In particular, it's a pain
when pulling chunks, because you've just pulled a chunk into your
current pane, and then your cursor is moved off this chunk. Now maybe
that's okay because you're done with it, but maybe you're not. You may
want to edit that chunk further, or (in three-way diff cases) you may
want to then push that chunk to the other side (i.e., in center, pull
from left, push to right).

If this doesn't end up working out, the change is trivial to revert.
However, for now this feels more correct to me.