diffmap: Call the wrong function, for GTK+'s insanity
Even though queuing a resize here should be totally fine, GTK+
completely loses it and complains about allocation issues. Given that
this is code that should basically never be necessary in our layout
(i.e., every time a scrollbar gets a size-allocate, we can basically
guarantee that DiffMap will as well) and given how annoying these
warnings are, we'll give in and just do the wrong thing.