filediff: Remove Shift-based scroll lock that GTK+ 3.18 broke
In GTK+ 3.18, default behaviour changed so that Shift + scroll scrolled
horizontally, instead of not having a binding. Since we bound this
to disable our vertical scroll lock, this broke that behaviour. However,
no one has complained at this point, so rather than find a new binding
I'm just going to remove it and see where we get to.