diffgrid: Ignore natural width when doing minimum width calculations
Ideally we'd be doing real calculations here, but we're not and this
works for now. The logic here is explained in the comment, but really
we're just trying to return a sane minimum width to avoid widgets
(realistically, just labels) that have a very large natural width but
support height-for-width geometry.

One user-visible bug caused by this is that when a long-highlighting
warning appears (which has a lot of text in a label), resizing panes
while it's visible chooses very large minimum sizes and breaks.