filediff, misc: Align syncpoints at top/bottom better (bgo#780608)
The issue here addressed by this change is that when we get scroll near
the top or bottom of a pane, and the corresponding pane that gets its
scroll set has more lines in the top-/bottom-most chunk, we won't ever
scroll the whole dependent chunk in to view.

For example, in a situation where we're scrolling to the bottom of a
comparison in the left pane, with the right pane being set by this
sync, and we have an end-of-file that looks like:

viewport  left   right
    |     x      x
    |     x      x
    |     x      x
    |     a      b
    _     EOF    b
                 b
                 EOF

the last two lines of b will often not be scrolled into view. The reason
here is that the way we decide how to align the right pane is to find
the middle of the left pane, figure out what chunk is there, find the
corresponding chunk in the right pane and then try to align those lines.
This means that the start and end of a file are aligned according to
whatever chunk is half way up the page. This is fine and good, except
that if e.g., the first `x` above is in the middle of the screen,
there's no way that you can scroll the left pane down far enough to be
able to see the bottom of the right pane.

The fix here is that we adjust the syncpoint that we use according to
where we are in the document. If we're showing the first half page, we
scale our syncpoint to be toward the top of the page; if we're showing
the last half page we scale the syncpoint towards the bottom. For most
normal documents, the majority of the document will have a syncpoint
of 0.5 (the middle of the screen) which is what we've always used.