filediff: Remove obsolete cursor handling for identical files case
This handling was added for the case when files are identical and so
there's no chunk to go to. Previously, we had cursor signals
disconnected at this point, so this clause existed just so that we would
have a correct line/column number in the status bar. With more recent
changes we now get our cursor handling triggered anyway, so this can go.