filediff: Dodge divide-by-zero in some edge cases
I'm not totally clear on when this happens, but in some cases (e.g.,
during slow file loading, when line revalidation is taking a while) we
occasionally get lines with a height of zero. Since this adjustment
exists mainly for smoother scrolling, it's easiest to just avoid the
zero rather than doing anything complicated.