filediff, dirdiff: Fix busted overlay scrolling change
The change in 1915de was supposed to just move overlay scrolling
disablement for dirdiff; instead it moved it for filediff and removed
it for dirdiff. This is the other part of the changeset, removing the
lagging code in filediff and actually disabling overlay scrolling in
dirdiff in the UI file.