diffgrid: Stop chaining up to avoid re-allocation loop (bgo#779883)
This commit is basically "I give up".

Because of GTK+'s use of the private GtkCSSGadget class in a bunch
of widgets, these widgets have become, as far as I can tell, effectively
impossible to subclass. When we hit issues with logging complaining
about missing allocation (despite no actual user-visible effects) the
advice was to always chain up when doing allocation. While this silenced
the warnings, it turns out that in some cases it leads to re-allocation
loops that never resolve. In our case, this looks like
DiffGrid.do_size_allocate() being called (from the main event loop)
forever.

I've spent way too long trying to bisect this, but what with behaviour
changes and slight differences in the symptoms, I've run out of any
motivation. As a result, I'm applying this crime of a patch that makes
the problem go away and reverts to our previous working behaviour.