dirdiff: Re-evaluate cursor-based state on model change (bgo#785859)
Because of the linear search we do to find the prev/next file to jump
to, we aggressively cache that particular bit of state so that it
doesn't need to be re-evaluated every time the cursor changes. However,
this can bite us when the model changes under us.

The solution in this patch is to simply force a state re-evaluation when
this happens, both on initial load and when individual items in the tree
change.

This also gets rid of the slightly odd allowance we had for forcing a
re-evaluation by manually calling the cursor-changed callback, in favour
of just using this re-evaluation toggle.