dirdiff: Fix old, old regression in initial cursor position
No idea when this broke, but because we just set the selection instead
of setting the cursor position, when you start up a folder comparison
you couldn't immediately navigate with Alt+Up/Down; you needed a cursor
move first.

There's no good reason for this, so just fix it up by setting the cursor
on initial load instead.