vc.git: Fix handling for added and partially staged files (bgo#785603)
The way Meld handles partially staged files is a little bit interesting.
We accumulate metadata across both the diff-to-index and the
diff-to-tree that we do, but treat the diff-to-tree as the 'correct'
status; this is consistent with us not really supporting staging area
viewing properly.

However, the way we were doing this was incorrect. We were uniquifying
entries (which should never have been non-unique anyway; probably a
left-over from a previous entry output format) and in doing so unsorted
them. Because we *implictly* prioritised the diff-to-tree status while
iterating over these now-unsorted results, everything went bad.

The easiest fix here is to stop uniquifying entries and, for the sake
of being explicit, break out the cached from the non-cached entries,
so that future generations can see how this behaviour is intended to
function.