vcview: Add missing state refresh (bgo#786629)
In non-flattened view we incorrectly missed a state refresh when a tree
row was updated; this fixes that.

Annoyingly, it's also an extremely aggressive refresh in terms of
invalidating the whole subtree, but there's sadly little that can be
done about that without a significant change to how we traverse and
update the tree model.