Fix support for auto-compare from the command line (bgo#785630)
This also (best I can recall) adds support for auto-compare for folder
comparisons, which I'm reasonably sure never actually worked.

The worst part of this patch is that it relies on the existing scheduler
queuing to ensure that the initial comparison has finished before
running the auto-compare code. This is "fine", except that it's just one
more thing to go wrong if/when I finally get rid of the scheduler.