vcview: Fix invalid tree iter crash with queued refreshes (bgo#787256)
The crash observed was caused by operations on the tree iter after it
has become invalid due to tree manipulation. Because we queue partial
refreshes to an idle loop when the tree is in non-flattened mode, we
could get in a situation where we'd queued multiple refreshes on
different valid-at-the-time tree iterators, and all of the iterators
after the first were invalidated by the refresh of the first.

The solution used here is to simple pass a path around to our
recursive tree construction, and to allow that helper to do the row
replacement in situations where it's required.

This is most easily triggerable with SVN, because we queue a refresh for
each folder/file that you add, though it is probably reproducable in
other cases as well.