meldbuffer: Non-existent files should be presented as writable
The use case here is that someone does:
    meld a i_dont_exist b
and in this scenario, the middle pane should be presented as being
editable, because merging to the middle and then saving it is probably
what the user wanted.

This scenario also applies to e.g.,
    meld a b c --output d
where `d` hasn't actually been created yet. You definitely expect the
pane that will be saved to `d` to be editable by default.