Use native fonts on win32 instead of FontConfig (bgo#785313)
The version 1.34 selected as pangocairo_requirement is the first version
with PangoCairo.FontMap.set_default behaviour identical to the version
Pango 1.40 which was used for testing.
It maybe works with previous versions too, but they are ancient.

Code assumes that version requirement is satisfied if import of both
Gtk and PangoCairo succeeds since Gtk 3.14 depends on Pango 1.36