vc.bzr: Fix py3k None ordering (bgo#788487)
This code uses None to indicate status flags that we don't care about.
We don't handle the more structured status of Bazaar, and instead just
chuck everything into a set and then have a few membership tests for
things we care about.

In Python 2, having Nones in that set was fine and they'd just never
match anything. That aspect is fine in Python 3 as well, but because we
use max() to figure out the "most meaningful" status for an entry,
Python 3 complains because of None ordering handling. Since we don't
ever actually care about the None entries here, the easiest fix is to
just remove that status.