Rework recent comparison API to return Gio.File instead of a URI
Previously this took a path, and the with the other Gio.File changes
changed to take a URI. While this is fine, there's no reason we can't
do the URI conversion on the recent files end and avoid doing it
in multiple places elsewhere.

As a bonus, this fixes a traceback when you tried to compare a file
with None (e.g., by using the New Comparison tab and selecting only one
file, leaving the second selector empty).