sourceview: Make inline-highlighted text show whitespace (bgo#589366)
The rationale here is that if your differences include whitespace, it's
probably best that you be able to see the whitespace characters. The
downside is that under normal GtkSourceView themes, the whitespace
indicators are quite low-contrast, so between that and the coloured
background in these chunks, the indicators are very difficult to see.

This is being included on a provisional basis until I decide whether
it's actually okay or whether we need to fix some stuff before it goes
in.