ui.bufferselectors: Add a TODO for fixing our duplicate UI files
There's a few issues with fixing this. The first and more serious is
that there doesn't seem to be a way to have the template of e.g.,
FilteredListSelector used for init_template() when subclassing. The
other issue is that I couldn't figure out a sane way to access objects
declared outside of the initialised template scope, such as the tree
model.