vc._vc: Limit the file information we ask for
We were asking for the full range of standard:: info, but didn't use
anything except the display name and whether it was a directory. Since
the standard namespace includes a *lot* of stuff, such as file type
sniffing, this was very wasteful.

On large version control hierarchies, this is good for a massive speed
improvement; I've seen up to 400%.