undo: Remove state assertions when clearing the undo sequence (ggo#152)
This assertion kind of makes sense, in that you *shouldn't* ever do
this, but on the other hand the clear method here should absolutely work
either way.

This is because there are actually cases where we apparently get a
user-action start without a user-action finish (e.g., in some file
loading error cases). While I'm tempted to argue that this is a bug in
the file load handling (in this case GtkSourceView), we also shouldn't
be crashing here for no good reason.