ui.statusbar: Don't update statusbar encoding directly
We need this to go through the signal logic, to the file comparison
checks, and finally through the buffer, where we'll change because of
the GObject property bind. If we don't do this then failed reloads
because of e.g., encoding errors, or users having unsaved changes will
mean that we're showing the wrong encoding in the status bar.