matchers.helpers: Discard queued items that haven't been picked up
In our worker model, we don't care even a little bit if some job that
gets queued by the main process gets dropped on the floor.

More importantly, if we *don't* do this, then quitting Meld while
highlighting jobs are still in the queue will absolutely deadlock the
process. I believe this is because the main process is waiting for
workers to read all of its items from the task queue before exiting,
but... the tools to debug this aren't great.