data/ui: Reword "revert" to "discard" in confirmation dialog
This is partially because we're reusing this now for confirmation on
encoding change, but also because it's actually a better description of
what we want to confirm. We don't want to confirm whether the user wants
to revert... we want to confirm whether they want to discard their
current changes. That's the destructive part of the operation.