filediff: Make unsaved change warning specific to reloaded files (#161)
Our "do you want to discard changes?" confirmation was written for the
revert-to-saved case, which reverts all panes at once. We're now reusing
it for the reload caused by encoding change, but it's still warning if
any file has been modified, even though it only reloads one file.

This change makes it so that the unsaved changes check takes a list of
panes that are being reloaded, and uses that in the encoding handler.