meldwindow: Fake out our progress spinner on Windows (#133)
On Windows, the spinner repaint is so slow (and on a high enough idle
priority) that it stops us making any comparison progress. This looks
like a Cairo problem, but either way it's not something we can really
fix in Meld, so we're doing the dumbest possible thing here and shimming
out the spinner.