sourceview: Don't try to handle a highlight on a non-realized sourceview
This can happen when a sourceview is created in the background (i.e., in
a non-focused tab), has a comparison loaded, and tries to go to the
first chunk. At that point, the go-to-chunk code queues a highlight
animation for the target chunk, but the sourceview containing that chunk
may not have been realized at all.