sourceview: Intercept and sanitise text-tags on paste
This is a different way of handling bgo#709580, with the change
triggered by #152.

The original problem here is that if you copy-paste within Meld, the
tags that we use to indicate an inline difference get pasted as well, by
GtkTextView/Buffer. This behaviour seems very difficult to opt out of.

The first pass at handling this hacked around the problem by hooking in
to tag apply and, if we were in an action group, skipping it. This
basically worked (but was fragile) until we hit an error situation where
we had unclosed action groups, at which point no tags were ever applied.

This second pass instead hooks in to the textview paste_clipboard vfunc
and requests and sets a text-only representation in the clipboard. This
effectively sanitises the clipboard contents and so... seems to work.