filediff: Work around GtkSourceView load failure (final part of #152)
When GtkSourceView's loader fails to load due to a unicode decode issue,
at least one of the failure cases doesn't actually end the in-progress
user action (see bgo#795387). Because we use user-actions to handle both
our undo stack and also our sensitivity setting (i.e., we disable some
things while you're in the middle of a user action) this breaks several
things across the Meld UI, basically putting us in a very weird state.

The workaround here is... not good, but it's the best we can do for now.
If/when the GtkSourceView side of this gets fixed, we'll end up with a
lot of GTK+ critical logs, but everything *should* continue to work.