filediff: Add a helper for showing user-criticial errors
The idea behind this decorator is that it should be used with functions
where failure is important to the user (and may not be obvious
otherwise). The only current use is on save, where it's possible to hit
a traceback and have the callback handling crash out, but other than a
log message or traceback on the command line... the user may not notice
anything.

This is made necessary by the GTK+/pygobject philosophy (which is
generally extremely good) of treating errors of this kind as non-fatal.
We also want it to not be fatal in this case, but... the user needs to
know if they tried to save a file and it didn't work.

On the other hand, if they tried to e.g., go to the next diff and we
messed something up and didn't go to the next diff, (a) they'd notice
and (b) showing them a modal dialog would be even more annoying than
the actual error.