sourceview: Fix pasting multi-byte encodings (#179)
For this API, pygobject isn't handling string length arguments for us,
so we had to do it ourselves. Shockingly, I got it very wrong. The API
requires a UTF-8 encoded string (which pygobject transparently encoded
for us) but the length arg has to be the bytestring length, not the
unicode codepoint length, which is what we were previously passing
along.