appveyor: Re-enable non-tag builds
Appveyor has some issues building tags if the tag isn't actually the
most recent commit on that branch. In our release workflow this is
theoretically fine, but because we're mirroring the repos across to
Github, in reality I've always pushed the post-release version bump
commit before the tag actually gets mirrored.