win32: Don't force the win32 pangocairo backend
The PangoCairo.Win32FontMap type used here to force the backend was supposed
to be private, but was exposed in the gir nevertheless. With the move of the
pango package in MSYS2 from autotools to meson the type is no longer
accessible and meld fails to start.

See https://github.com/Alexpux/MINGW-packages/issues/3610 for the error report

Another supported way to change the backend would be to do:
    os.environ['PANGOCAIRO_BACKEND'] = 'win32'
before importing gtk, but with fontconfig 2.13 the initial cache creation
time was greatly reduced (https://github.com/Alexpux/MINGW-packages/pull/3480)
and the workaround can be removed.