meldapp: Don't try to open a blank comparison tab (#193)
We have existing code in meldapp that handles this 90% fine. However,
if the user does something like always using `--diff` (which is
particularly understandable for e.g., tooling) then invocations like:
    meld -n --diff a b
will result in the new tab not being focused on open, because it's not
first in the list of comparisons we're opening.