maint: Add a command for appveyor build triggers
This exists partially for testing, but mostly because the current
fragmentation between github and gitlab messes with out build pipeline,
and so we don't actually get a simple build off a tag. This makes it
easy to rebuild a tag after the fact so that we can get a usable binary.