Use external help in frozen installation, like uninstalled (#197)
There were two different UNINSTALLED_* flags for
uninstalled/frozen installations. Every flag had a single use.

In #197 it turns out that showing offline help in frozen installation
can fail since frozen installation is quite "portable"
so it can be problematic to rely on system tool for showing help.
(actually I'm not completely sure how freezing is typically
used on non-windows, but non-windows freezing doesn't look popular)

So simple fix of #197 is to merge flags, because the single case
that was different (trying to sho help in frozen state) is the bug case.
So it looks that single flag indicating that meld can assume that
it's data is installed in "global" system prefixes is enough.