filediff: Fix change deletion shortcut for non-keypad deletes
This was broken by 30bd3231, because you can't actually expect multiple
accelerators to work.

The change here is kind of unpleasant. Really, it would be nice if
this was a generic helper that ran on focus, probably as part of MeldDoc
handling... but that's very overkill for this one accelerator.