filediff: Support drag-and-drop directly on to textviews (#177)
The behaviour here is hopefully fairly intuitive. If the user drops a
single URI on a file pane, we will load it into that pane. If they drop
several URIs, we'll check to see whether they've dropped the right
number to load all panes, and if so then load them all.

This doesn't handle correctly prelighting of drop zones to e.g., show
the user that they can't drop two files on to a three-way comparison.
This omission is partly due to laziness, but mostly because I don't
think it's a big deal and the drag-motion signal is a bit wild given
that we'd have to coordinate our handling with GtkTextView and also need
to handle the request data dance with drag-data-received.