dirdiff: Fix state filter initialisation logic
The callback for setting the active state of the filter buttons entirely
replaces the list we're iterating over. This leads to some extremely
odd interactions, and the UI state desyncs from what we think we have
stored in settings. One better fix here might be to have the filter
callback sync individual items or something instead of recreating the
whole list, but... this is less invasive.

It's somewhat surprising that this hasn't bitten us before.