dirdiff: Refactor error row adding for cleaner defaults
The only reason these defaults aren't in the base `tree` module is that
vcview doesn't have the columns. I've tried using column defaults
elsewhere, but `GtkTreeStore` always seems to want to set column values
to a zero/null-esque empty state, so setting it to -1 by default doesn't
seem to be easily doable.