filediff: Remove non-functional encoding fallback check (#239)
The intention of this check is to allow users to fall back to using
UTF-8 if, for example, they loaded a file that was auto-detected as
ASCII and then typed in anything outside of that charset.

Clearly this is a niche use case because it hasn't worked since 3.15.2
(because we never use the assigned-to `source_encoding`), and no one has
complained about the fallback; the bug here is that the check itself
doesn't work any more.

In 3.19.x this is fixable by the user with the new encoding selector, so
realistically that's probably going to be the answer here.