filediff: Replace broken pre-save encoding check (#239)
See c7bbe78d for the removal reasoning.

Here we're just giving the user a somewhat better error message. Ideally
we'd replicate the previous offer of "hey why don't you save as UTF-8
instead?", but the save lifecycle makes this difficult here. If this was
an often-hit check then we'd do it, but... this code hasn't worked in
over three years and people have only ever complained about bugs in the
encoding check, not the save-as-UTF-8 bit.