windowstate: Avoid spurious dconf writes on allocate events (#247)
Currently we write the window state a *lot*, because we constantly get
spammed with `size-allocate` signals when nothing has changed (e.g.,
we've scrolled a comparison). The window size is typically unchanged,
so here we add the simplest possible check before assigning to our
bound properties. In my testing, this completely eliminates the spurious
writes.

There's also the dconf part of this where it's constantly re-writing
values that haven't actually changed; see GNOME/dconf#40 for the
discussion about that.