filediff: Rework the keymask-focus handling to be purely textview-local
The bug that this code was addressing goes like this:
 * Hold down Shift to trigger delete mode
 * Switch window focus to another application
 * File comparison actions are stuck in delete mode even after
   refocusing, until Shift is pressed again.

The solution that we had addressed this directly by hooking in to the
window focus change. This worked (except for e.g., switching the focus
widget within Meld itself...) but was complete overkill. The new fix
here takes the significantly saner approach of just always resetting
the keymask when you unfocus a text view.