preferences: Migrate list widgets to templates and resource loading
This is a big commit because the old separation between the
ListWidget class and the actually-used subclasses was kind of wild.

Now, EditableListWidget is much simpler and exists mostly to handle
sensitivity setting and simple treemodel manipulation. Everything else
lives on the actual list widgets themselves, which are now real widgets
that do nice, normal GObject property things.