meldwindow, ui.notebook: Update containers to handle actual GtkWidgets
MeldWindow currently expects to create, and MeldNotebook to contain,
just gnomeglade.Component subclasses. Since the idea of moving to
template loading is to make things more normal, Component is going away.

This commit just makes it so that these places can accept an actual
widget instead of a special Python object that wraps the widget in a
known way.