dirdiff: Fix menu focus issues by removing a *lot* of code
The actual reason I went to change this was that if you tried to use the
menu bar items, the main actions (i.e., compare, copy, delete) were all
insensitive. This happens because we unset their sensitivity on treeview
focus loss. I suspect that the intention here was to keep consistent
action state, but in reality we can just invoke the action based on the
last focus pane, which is what the user expects.

When doing this, I found that the focus handler block/unblock around
popup handling was doing nothing useful, and then that the popup menu
handling was actually slightly confused with the custom filter menu
code, so... now it's mostly deleted.