dirdiff: Add Enter/Return as an accelerator for Compare (#262)
The intention here was for this to essentially be a documentation
change, i.e., so that users can see that the shortcut exists. In
reality, this is a behaviour change when hitting enter on a folder row.

Previously, this would expand/collapse the row because it was a treeview
row activation signal and that seemed like a reasonable thing to do.
This is still what will happen on actual row activation (i.e., double-
clicking a folder row), but now hitting enter will activate the Compare
action and will create a new folder comparison when activating a folder
row.

I don't think either of these options are obviously better than the
other, so I'm going to go with this new behaviour for now.