filediff: Move users of check-save workflow to confirm-discard instead
The benefit here is that while state in check-save is async (because we
use the async saving of GtkSource.FileSaver), confirm-discard has no
state and is a straight up modal confirm with no complications.

The places we're now doing this is when the user changes the file
selector, and when they drag-n-drop a new file on to the pane. In both
cases, I think these aren't situations where the user is likely to
actually have accidentally forgotten to save (since they're extremely
deliberate manipulations of the current comparison), so the simple
confirm seems like a reasonable complexity trade-off.