Move notebook label UI to be full-width and hidden with only one tab
This is the UI form used by basically every modern GTK+/GNOME-y MDI UI
that I can find. It has its drawbacks (mainly that it's different to the
browser MDI UI, which everyone uses), but it circumvents other issues
such as us having very long labels in very small tabs.

This fixes (even if indirectly) #21 and #263.