settings: Support one-way gsettings bindings and use for source view
When these settings are changed on the sourceview, we absolutely do not
want that to be propagated back to the gsettings store. Currently these
settings *can't* be changed, so this isn't a problem. However, we're
about to add some UI for that, so this is preparation.

Also, probably most of our gsettings bindings should actually be one-
way, but they do require actual thought.