sourceview: Expose a Boolean-ised text wrapping property
This exists so that we have a property that can be bound to a
PropertyAction for use in a popover menu toggle.

I experimented with *not* doing this and exposing three radio buttons to
match our three possible wrapping states, but it felt extremely clunky
in the UI and was hard to label.

The logic here for deciding what "enable wrapping" means here looks
pretty odd, but it comes down to:

 * if we have some wrapping setting (i.e., wrap at characters) by
   default then respect that when re-enabling the wrapping; and
 * if we don't wrap by default then just use word wrapping.

since this was the behaviour that felt least-wrong to me.